requirements.txt
Flask
flask-cors
gunicorn
requests
python-dotenv
google-generativeai
